+++ RELEASE_NOTES.md
@@ -235,5 +235,6 @@
 ### 3.1.0 - Unreleased
 
 * Added `TimeOnly` GraphQL type
+* Added `map` function for Relay `Connection` and `Edge` types
 * Excluded `GraphQLWebSocketMiddleware` from exception stack trace if request not a Web Socket
 * Changed `GraphQLOptionsDefaults.WebSocketConnectionInitTimeoutInMs` const type to `double`